# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tratamientos": {
            "get": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Listar todos los tratamientos",
                "operationId": "531c467fa1225b865227c3189bd19f5f",
                "responses": {
                    "200": {
                        "description": "Tratamientos encontrados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseTratamientoList"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron tratamientos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontraron tratamientos"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Crear un nuevo tratamiento",
                "operationId": "d66ebf8168340757953ee6187b57626b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_tratamiento",
                                    "descripcion",
                                    "duracion",
                                    "precio"
                                ],
                                "properties": {
                                    "id_tratamiento": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Tratamiento facial"
                                    },
                                    "duracion": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tratamiento creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseTratamientoSingle"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidacion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tratamientos/{id}": {
            "get": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Obtener un tratamiento por ID",
                "operationId": "38277f87680261905ab7b8dde40808a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tratamiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tratamiento encontrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseTratamientoSingle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tratamiento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tratamiento no encontrado"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Actualizar un tratamiento por ID",
                "operationId": "70fcd13f536851096c46b104d21bf281",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tratamiento a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_tratamiento": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Tratamiento facial actualizado"
                                    },
                                    "duracion": {
                                        "type": "integer",
                                        "example": 75
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1300
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tratamiento actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseTratamientoSingle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tratamiento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tratamiento no encontrado"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidacion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tratamientos"
                ],
                "summary": "Eliminar un tratamiento por ID",
                "operationId": "f66cc20c9860cf5dd974c941430e16ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tratamiento a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tratamiento eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tratamiento eliminado exitosamente"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tratamiento no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tratamiento no encontrado"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tratamiento": {
                "properties": {
                    "id_tratamiento": {
                        "type": "integer",
                        "example": 1
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Tratamiento facial"
                    },
                    "duracion": {
                        "type": "integer",
                        "example": 60
                    },
                    "precio": {
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    }
                },
                "type": "object"
            },
            "ErrorValidacion": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error en la validación de datos"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "property": null
                        }
                    },
                    "success": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ResponseTratamientoList": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Tratamientos encontrados correctamente"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tratamiento"
                        }
                    },
                    "success": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "ResponseTratamientoSingle": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Tratamiento encontrado correctamente"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Tratamiento"
                    },
                    "success": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}